+++ src/components/ChartOfAccounts/index.tsx
@@ -1,6 +1,1 @@
-export {
-  ChartOfAccounts,
-  ChartOfAccountsContext,
-  LedgerAccountsContext,
-  View,
-} from './ChartOfAccounts'
+export { ChartOfAccounts, View } from './ChartOfAccounts'
