+++ src/components/ChartOfAccountsTable/ChartOfAccountsTable.tsx
@@ -1,7 +1,8 @@
 import React, { RefObject, useContext, useState } from 'react'
+import { ChartOfAccountsContext } from '../../contexts/ChartOfAccountsContext'
 import DownloadCloud from '../../icons/DownloadCloud'
 import { Button, ButtonVariant } from '../Button'
-import { ChartOfAccountsContext, View } from '../ChartOfAccounts'
+import { View } from '../ChartOfAccounts/ChartOfAccounts'
 import { ChartOfAccountsDatePicker } from '../ChartOfAccountsDatePicker'
 import { ChartOfAccountsRow } from '../ChartOfAccountsRow'
 import { ChartOfAccountsSidebar } from '../ChartOfAccountsSidebar'
