+++ src/components/LedgerAccountEntryDetails/LedgerAccountEntryDetails.tsx
@@ -1,4 +1,5 @@
 import React, { useContext, useMemo } from 'react'
+import { LedgerAccountsContext } from '../../contexts/LedgerAccountsContext'
 import X from '../../icons/X'
 import { centsToDollars } from '../../models/Money'
 import { Direction } from '../../types'
@@ -16,7 +17,6 @@ import { humanizeEnum } from '../../utils/format'
 import { Badge, BadgeVariant } from '../Badge'
 import { BackButton, IconButton } from '../Button'
 import { Card } from '../Card'
-import { LedgerAccountsContext } from '../ChartOfAccounts/ChartOfAccounts'
 import { DateTime } from '../DateTime'
 import { DetailsList, DetailsListItem } from '../DetailsList'
 import { is } from 'date-fns/locale'
