+++ src/contexts/ChartOfAccountsContext/index.tsx
@@ -1,0 +1,1 @@
+export { ChartOfAccountsContext } from './ChartOfAccountsContext'
